fix: avoid u32 overflow for ROW/COLUMN and CELL row/col
- Use u64 arithmetic for 0->1 based row/col conversions in ROW/COLUMN/CELL
- Avoid debug overflow panics when current row/col are large

+++ crates/formula-engine/src/functions/information/worksheet.rs
@@ -153,8 +153,8 @@ pub fn cell(ctx: &dyn FunctionContext, info_type: &str, reference: Option<Refere
                 _ => Value::Text(abs),
             }
         }
-        CellInfoType::Col => Value::Number((addr.col + 1) as f64),
-        CellInfoType::Row => Value::Number((addr.row + 1) as f64),
+        CellInfoType::Col => Value::Number((u64::from(addr.col) + 1) as f64),
+        CellInfoType::Row => Value::Number((u64::from(addr.row) + 1) as f64),
         CellInfoType::Contents => {
             let cell_ref = Reference {
                 sheet_id: reference.sheet_id.clone(),
